+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `@microsoft/eslint-plugin-sdl` eslint plugin
+
 ## [1.0.0] - 2023-04-14
 
 ### Added
 
 - unit tests for allowing an element to be used twice in a conditional predicate
 - support for NodeJS 18
-- `@microsoft/eslint-plugin-sdl` eslint plugin
 
 ### Removed