+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-05-08
+
 ### Changed
 
 - **[BREAKING]** `submissionService.replaceElementValues()` to `submissionService.replaceInjectablesWithElementValues()`